Remove unused scripts
